// packages/beecli/sdk/beetrade-v2/models/CreateStrategyRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Beetrade API
 * API documentation for Beetrade
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CreateStrategyRequest
 */
export interface CreateStrategyRequest {
    /**
     * 
     * @type {{ [key: string]: any; }}
     * @memberof CreateStrategyRequest
     */
    data: { [key: string]: any; };
    /**
     * 
     * @type {string}
     * @memberof CreateStrategyRequest
     */
    description?: string;
    /**
     * 
     * @type {Array<string>}
     * @memberof CreateStrategyRequest
     */
    indicators?: Array<string>;
    /**
     * 
     * @type {string}
     * @memberof CreateStrategyRequest
     */
    name: string;
}

/**
 * Check if a given object implements the CreateStrategyRequest interface.
 */
export function instanceOfCreateStrategyRequest(value: object): value is CreateStrategyRequest {
    if (!('data' in value) || value['data'] === undefined) return false;
    if (!('name' in value) || value['name'] === undefined) return false;
    return true;
}

export function CreateStrategyRequestFromJSON(json: any): CreateStrategyRequest {
    return CreateStrategyRequestFromJSONTyped(json, false);
}

export function CreateStrategyRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreateStrategyRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'data': json['data'],
        'description': json['description'] == null ? undefined : json['description'],
        'indicators': json['indicators'] == null ? undefined : json['indicators'],
        'name': json['name'],
    };
}

export function CreateStrategyRequestToJSON(json: any): CreateStrategyRequest {
    return CreateStrategyRequestToJSONTyped(json, false);
}

export function CreateStrategyRequestToJSONTyped(value?: CreateStrategyRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'data': value['data'],
        'description': value['description'],
        'indicators': value['indicators'],
        'name': value['name'],
    };
}