// packages/beecli/sdk/beetrade-v2/models/UpdateStrategyRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Beetrade API
 * API documentation for Beetrade
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface UpdateStrategyRequest
 */
export interface UpdateStrategyRequest {
    /**
     * 
     * @type {{ [key: string]: any; }}
     * @memberof UpdateStrategyRequest
     */
    data?: { [key: string]: any; };
    /**
     * 
     * @type {string}
     * @memberof UpdateStrategyRequest
     */
    description?: string;
    /**
     * 
     * @type {Array<string>}
     * @memberof UpdateStrategyRequest
     */
    indicators?: Array<string>;
    /**
     * 
     * @type {string}
     * @memberof UpdateStrategyRequest
     */
    name?: string;
}

/**
 * Check if a given object implements the UpdateStrategyRequest interface.
 */
export function instanceOfUpdateStrategyRequest(value: object): value is UpdateStrategyRequest {
    return true;
}

export function UpdateStrategyRequestFromJSON(json: any): UpdateStrategyRequest {
    return UpdateStrategyRequestFromJSONTyped(json, false);
}

export function UpdateStrategyRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): UpdateStrategyRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'data': json['data'] == null ? undefined : json['data'],
        'description': json['description'] == null ? undefined : json['description'],
        'indicators': json['indicators'] == null ? undefined : json['indicators'],
        'name': json['name'] == null ? undefined : json['name'],
    };
}

export function UpdateStrategyRequestToJSON(json: any): UpdateStrategyRequest {
    return UpdateStrategyRequestToJSONTyped(json, false);
}

export function UpdateStrategyRequestToJSONTyped(value?: UpdateStrategyRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'data': value['data'],
        'description': value['description'],
        'indicators': value['indicators'],
        'name': value['name'],
    };
}