// packages/beecli/sdk/beetrade-v2/models/CreateBacktestRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Beetrade API
 * API documentation for Beetrade
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CreateBacktestRequest
 */
export interface CreateBacktestRequest {
    /**
     * 
     * @type {number}
     * @memberof CreateBacktestRequest
     */
    brokerageId: number;
    /**
     * 
     * @type {number}
     * @memberof CreateBacktestRequest
     */
    cash: number;
    /**
     * 
     * @type {string}
     * @memberof CreateBacktestRequest
     */
    endDate: string;
    /**
     * 
     * @type {string}
     * @memberof CreateBacktestRequest
     */
    startDate: string;
}

/**
 * Check if a given object implements the CreateBacktestRequest interface.
 */
export function instanceOfCreateBacktestRequest(value: object): value is CreateBacktestRequest {
    if (!('brokerageId' in value) || value['brokerageId'] === undefined) return false;
    if (!('cash' in value) || value['cash'] === undefined) return false;
    if (!('endDate' in value) || value['endDate'] === undefined) return false;
    if (!('startDate' in value) || value['startDate'] === undefined) return false;
    return true;
}

export function CreateBacktestRequestFromJSON(json: any): CreateBacktestRequest {
    return CreateBacktestRequestFromJSONTyped(json, false);
}

export function CreateBacktestRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreateBacktestRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'brokerageId': json['brokerageId'],
        'cash': json['cash'],
        'endDate': json['endDate'],
        'startDate': json['startDate'],
    };
}

export function CreateBacktestRequestToJSON(json: any): CreateBacktestRequest {
    return CreateBacktestRequestToJSONTyped(json, false);
}

export function CreateBacktestRequestToJSONTyped(value?: CreateBacktestRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'brokerageId': value['brokerageId'],
        'cash': value['cash'],
        'endDate': value['endDate'],
        'startDate': value['startDate'],
    };
}