// packages/beecli/sdk/beetrade-v2/models/CreatePaperTradingRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Beetrade API
 * API documentation for Beetrade
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CreatePaperTradingRequest
 */
export interface CreatePaperTradingRequest {
    /**
     * 
     * @type {number}
     * @memberof CreatePaperTradingRequest
     */
    brokerageId: number;
    /**
     * 
     * @type {number}
     * @memberof CreatePaperTradingRequest
     */
    cash: number;
}

/**
 * Check if a given object implements the CreatePaperTradingRequest interface.
 */
export function instanceOfCreatePaperTradingRequest(value: object): value is CreatePaperTradingRequest {
    if (!('brokerageId' in value) || value['brokerageId'] === undefined) return false;
    if (!('cash' in value) || value['cash'] === undefined) return false;
    return true;
}

export function CreatePaperTradingRequestFromJSON(json: any): CreatePaperTradingRequest {
    return CreatePaperTradingRequestFromJSONTyped(json, false);
}

export function CreatePaperTradingRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreatePaperTradingRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'brokerageId': json['brokerageId'],
        'cash': json['cash'],
    };
}

export function CreatePaperTradingRequestToJSON(json: any): CreatePaperTradingRequest {
    return CreatePaperTradingRequestToJSONTyped(json, false);
}

export function CreatePaperTradingRequestToJSONTyped(value?: CreatePaperTradingRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'brokerageId': value['brokerageId'],
        'cash': value['cash'],
    };
}