// packages/beecli/sdk/beetrade-v2/models/CreateLiveTradingRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Beetrade API
 * API documentation for Beetrade
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CreateLiveTradingRequest
 */
export interface CreateLiveTradingRequest {
    /**
     * 
     * @type {number}
     * @memberof CreateLiveTradingRequest
     */
    brokerageAccountId: number;
}

/**
 * Check if a given object implements the CreateLiveTradingRequest interface.
 */
export function instanceOfCreateLiveTradingRequest(value: object): value is CreateLiveTradingRequest {
    if (!('brokerageAccountId' in value) || value['brokerageAccountId'] === undefined) return false;
    return true;
}

export function CreateLiveTradingRequestFromJSON(json: any): CreateLiveTradingRequest {
    return CreateLiveTradingRequestFromJSONTyped(json, false);
}

export function CreateLiveTradingRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreateLiveTradingRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'brokerageAccountId': json['brokerageAccountId'],
    };
}

export function CreateLiveTradingRequestToJSON(json: any): CreateLiveTradingRequest {
    return CreateLiveTradingRequestToJSONTyped(json, false);
}

export function CreateLiveTradingRequestToJSONTyped(value?: CreateLiveTradingRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'brokerageAccountId': value['brokerageAccountId'],
    };
}