// packages/beecli/sdk/beetrade-v2/models/ScheduleStrategyRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Beetrade API
 * API documentation for Beetrade
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ScheduleStrategyRequest
 */
export interface ScheduleStrategyRequest {
    /**
     * 
     * @type {string}
     * @memberof ScheduleStrategyRequest
     */
    cronExpression: string;
}

/**
 * Check if a given object implements the ScheduleStrategyRequest interface.
 */
export function instanceOfScheduleStrategyRequest(value: object): value is ScheduleStrategyRequest {
    if (!('cronExpression' in value) || value['cronExpression'] === undefined) return false;
    return true;
}

export function ScheduleStrategyRequestFromJSON(json: any): ScheduleStrategyRequest {
    return ScheduleStrategyRequestFromJSONTyped(json, false);
}

export function ScheduleStrategyRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): ScheduleStrategyRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'cronExpression': json['cronExpression'],
    };
}

export function ScheduleStrategyRequestToJSON(json: any): ScheduleStrategyRequest {
    return ScheduleStrategyRequestToJSONTyped(json, false);
}

export function ScheduleStrategyRequestToJSONTyped(value?: ScheduleStrategyRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'cronExpression': value['cronExpression'],
    };
}