// packages/beecli/sdk/beetrade-v2/models/CreateAlertRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Beetrade API
 * API documentation for Beetrade
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { AlertType } from './AlertType';
import {
    AlertTypeFromJSON,
    AlertTypeFromJSONTyped,
    AlertTypeToJSON,
    AlertTypeToJSONTyped,
} from './AlertType';
import type { AlertCondition } from './AlertCondition';
import {
    AlertConditionFromJSON,
    AlertConditionFromJSONTyped,
    AlertConditionToJSON,
    AlertConditionToJSONTyped,
} from './AlertCondition';

/**
 * 
 * @export
 * @interface CreateAlertRequest
 */
export interface CreateAlertRequest {
    /**
     * 
     * @type {CreateAlertRequestActionEnum}
     * @memberof CreateAlertRequest
     */
    action: CreateAlertRequestActionEnum;
    /**
     * 
     * @type {{ [key: string]: any; }}
     * @memberof CreateAlertRequest
     */
    actionParameters?: { [key: string]: any; };
    /**
     * 
     * @type {AlertCondition}
     * @memberof CreateAlertRequest
     */
    condition: AlertCondition;
    /**
     * 
     * @type {string}
     * @memberof CreateAlertRequest
     */
    name: string;
    /**
     * 
     * @type {number}
     * @memberof CreateAlertRequest
     */
    thresholdValue: number;
    /**
     * 
     * @type {AlertType}
     * @memberof CreateAlertRequest
     */
    type: AlertType;
}

/**
* @export
* @enum {string}
*/
export enum CreateAlertRequestActionEnum {
    StopAlert = 'STOP_ALERT',
    NotifyEmail = 'NOTIFY_EMAIL',
    NotifyTelegram = 'NOTIFY_TELEGRAM',
    SendWebhook = 'SEND_WEBHOOK'
}


/**
 * Check if a given object implements the CreateAlertRequest interface.
 */
export function instanceOfCreateAlertRequest(value: object): value is CreateAlertRequest {
    if (!('action' in value) || value['action'] === undefined) return false;
    if (!('condition' in value) || value['condition'] === undefined) return false;
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('thresholdValue' in value) || value['thresholdValue'] === undefined) return false;
    if (!('type' in value) || value['type'] === undefined) return false;
    return true;
}

export function CreateAlertRequestFromJSON(json: any): CreateAlertRequest {
    return CreateAlertRequestFromJSONTyped(json, false);
}

export function CreateAlertRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreateAlertRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'action': json['action'],
        'actionParameters': json['actionParameters'] == null ? undefined : json['actionParameters'],
        'condition': AlertConditionFromJSON(json['condition']),
        'name': json['name'],
        'thresholdValue': json['thresholdValue'],
        'type': AlertTypeFromJSON(json['type']),
    };
}

export function CreateAlertRequestToJSON(json: any): CreateAlertRequest {
    return CreateAlertRequestToJSONTyped(json, false);
}

export function CreateAlertRequestToJSONTyped(value?: CreateAlertRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'action': value['action'],
        'actionParameters': value['actionParameters'],
        'condition': AlertConditionToJSON(value['condition']),
        'name': value['name'],
        'thresholdValue': value['thresholdValue'],
        'type': AlertTypeToJSON(value['type']),
    };
}